# Force to fp16 models to fp32 if `XLA_USE_FP16` is already set.