Temporarily fix: #6348 

This PR forces the model's precision to be fp32 whenever we are running the benchmark with XLA:CUDA, and the ATen tensor will be downcasted to fp16 or bf16 (i.e. `XLA_USE_FP16` or `XLA_USE_BF16` are set).

```python
# Problem of having fp16 tensors with XLA_USE_FP16 set.
device = xm.xla_device()

m = torch.nn.ReLU().to(device).half()
input = torch.randn(100, device=device, dtype=dtype)

output = m(input)

print(input)
# tensor((100,), dtype=torch.float16, device=xla:0)
print(output)
# tensor((100,), dtype=torch.float32, device=xla:0)
```

The better way to do it is to:
- Wait for https://github.com/pytorch/pytorch/pull/117167 to land
- Not set `XLA_USE_FP16`

cc @miladm @JackCaoG