@golechwierowicz 2024-01-26
What's the timeline for landing the proper fix?

The entire logic of casting/downcasting seems really wobbly. From our side the only thing we care about is whether kernels execute for the same precision for every benchmarking configuration of the single model.