@ysiraichi 2024-01-26
The proper fix is being worked on the PyTorch side: https://github.com/pytorch/pytorch/pull/117167.

> From our side the only thing we care about is whether kernels execute for the same precision for every benchmarking configuration of the single model

That's what this PR is trying to accomplish. Specifically, there are 2 problems:

- We can't rely on `model.to(precision)`, because of https://github.com/pytorch/pytorch/issues/115792 (being fixed by the PR mentioned above)
- `XLA_USE_FP16` in combination with `model.to(torch.float16)` returns upcasted results, which breaks some PyTorch operations' type-check

The solution proposed by this PR is:

- Keep setting `XLA_USE_FP16`, which runs the computation on the desired `fp16` precision
- Force the tensors' dtype to be `fp32`, which will make PyTorch type checking to pass