@golechwierowicz 2024-01-26
I understand the problem, and the solution. My point is: is it better to wait for Meta fix, or proceed with this PR. 

The reason I am hesitant with the latter is because precision mismatches gave us a major headache, and we would like to avoid them if possible. Any changes to the precision logic, especially where additional complexity is introduced, are dangerous.

That being said, I'd be more confident to proceed with this PR if I saw a some form of a test.
Could you dump kernel profiles for e.g. hf_Bert, and resnet50 with this diff so we could take a look at kernels?