@ysiraichi 2024-01-26
> is it better to wait for Meta fix, or proceed with this PR

Good question. That depends on how close they are to landing that PR (which, I guess is what you asked, at first). I'm not seeing reviews there yet. So, it could take a while. The "good" thing about this PR is that it is bringing back the old behavior (before #6296) we had for these models.

> Could you dump kernel profiles for e.g. hf_Bert, and resnet50 with this diff so we could take a look at kernels?

Yeah, sure.