@ysiraichi 2024-01-29
@golechwierowicz

Here are the results. They seem to be similar enough. The table below refers to the **last** (since it does not account for tracing+compilation) of the total CUDA timings:

| | Before | After |
|----------|-------------|---------------|
| hf_Bert | 12.824ms | 13.076ms |
| resnet50 | 7.204ms | 7.187ms |
| vgg16 | 2.099ms | 2.093ms |

_(`vgg16` was added just as another sample for this comparison)_

```bash
python xla/benchmarks/experiment_runner.py \
    --suite-name torchbench --accelerator cuda --repeat 8 --test eval --print-subprocess --no-resume \
    --xla PJRT --dynamo openxla --dump-pytorch-profiles
```

-----

I'm sending the raw profiles externally, since it goes over the character limit of a GH comment.